Rename Gamepad.* to Device.*.
