+++ pcsx2/PAD/Linux/Device.h
