+++ pcsx2/PAD/Linux/ini.cpp
@@ -14,7 +14,7 @@
  */
 
 #include "Global.h"
-#include "GamePad.h"
+#include "Device.h"
 #include "keyboard.h"
 #include "AppConfig.h"
