+++ pcsx2/PAD/Linux/state_management.cpp
@@ -14,7 +14,7 @@
  */
 
 #include "state_management.h"
-#include "GamePad.h"
+#include "Device.h"
 
 // Typical packet response on the bus
 static const u8 ConfigExit[7] = {0x5A, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00};
