+++ pcsx2/PAD/Linux/wx_dialog/dialog.h
@@ -30,7 +30,7 @@
 
 #include <sstream>
 
-#include "../GamePad.h"
+#include "../Device.h"
 #include "../keyboard.h"
 #include "../Global.h"
 #include "opPanel.h"
